Revert "add and test player movement sync"
This reverts commit e86f1bcebc7e73aadd5557462318a0107fe1e097.

+++ Assets/_Scripts/Network/NetworkMessages/NetworkMessage.cs
@@ -18,7 +18,7 @@ public abstract class ClientMessage : NetworkMessage
     [JsonProperty("clientId")]
     public string SenderId;
     [JsonProperty("clientSendTime")]
-    public long ClientSendTime;
+    public double ClientSendTime;
     [JsonProperty("clientTick")]
     public int ClientTick;
     protected ClientMessage(short messageType) : base(messageType) { ClientSendTime = TimeSyncUtils.GetUnixTimeMilliseconds(); }
@@ -29,7 +29,7 @@ public abstract class ServerMessage : NetworkMessage
     [JsonProperty("clientId")]
     public string ReceiverId;
     [JsonProperty("serverSendTime")]
-    public long ServerSendTime;
+    public double ServerSendTime;
     [JsonProperty("serverTick")]
     public int ServerTick;
     protected ServerMessage(short messageType) : base(messageType) { }
